src/requirements.txt
# requirements.txt
streamlit_option_menu==0.3.13